First Shot
This is the original doc that I put together.  Originally the application pulled from an external excel document as the data store so that multiple users could be in the main application simultaneously.  When the user would save, it would unlock the data document and update the document with changes/additions made.  

I've got two options to make this work again.  1-have the application create the data file if one does not already exist.  2-put the data back in the existing Excel document and limit the use to one-user-at-a-time.  

If you want to tweak around with it, the passwords are "VVBS".